+++ src/openssl_wrap.cpp
@@ -1271,11 +1271,6 @@ void _EVP_PKEY_CTX_set_rsa_oaep_label(EVP_PKEY_CTX *ctx, unsigned char *l, int l
     OpensslCallIsNonNegative::callChecked(lib::OpenSSLLib::SSL_EVP_PKEY_CTX_set_rsa_oaep_label, ctx, l, llen);
 }
 
-int _EVP_PKEY_CTX_get_rsa_oaep_label(EVP_PKEY_CTX *ctx, unsigned char *l)
-{
-    return OpensslCallIsNonNegative::callChecked(lib::OpenSSLLib::SSL_EVP_PKEY_CTX_get_rsa_oaep_label, ctx, l);
-}
-
 int _RSA_size(const RSA *r)
 {
     return lib::OpenSSLLib::SSL_RSA_size(r);