+++ tests/unit/openssl_lib_mock.cpp
@@ -833,10 +833,6 @@ int OpenSSLLib::SSL_EVP_PKEY_CTX_set_rsa_oaep_label(EVP_PKEY_CTX *ctx, unsigned 
     return OpenSSLLibMockManager::getMockInterface().SSL_EVP_PKEY_CTX_set_rsa_oaep_label(ctx,
                                                                                          l, llen);
 }
-int OpenSSLLib::SSL_EVP_PKEY_CTX_get_rsa_oaep_label(EVP_PKEY_CTX *ctx, unsigned char *l) noexcept
-{
-    return OpenSSLLibMockManager::getMockInterface().SSL_EVP_PKEY_CTX_get_rsa_oaep_label(ctx, l);
-}
 int OpenSSLLib::SSL_RSA_size(const RSA *r) noexcept
 {
     return OpenSSLLibMockManager::getMockInterface().SSL_RSA_size(r);