+++ CHANGELOG.md
@@ -6,13 +6,22 @@ All notable changes to this project will be documented in this file.
 
 ## Changed
 
+## Fixed
+
+## Added
+
+# Release 4.1.0
+
+Support for AES-CMAC has been added (see `mac-example.cpp` for sample usage),
+as well as some minor changes listed below.
+
+## Changed
+
 * Remove wrapper `openssl::_EVP_PKEY_CTX_get_rsa_oaep_label`. This is
   technically an ABI break, but since the wrappers are not considered part of
   the public API, we do not bump the SOVERSION for this.
 * Improve error message in MoCOCrWException that is thrown in case of invalid
   signature validation.
-
-## Fixed
 
 ## Added